+++ notify-stack/src/main/java/com/dwarfeng/notify/stack/exception/RouterMakeException.java
@@ -8,20 +8,24 @@ package com.dwarfeng.notify.stack.exception;
  */
 public class RouterMakeException extends RouterException {
 
-    private static final long serialVersionUID = -985767174276194192L;
+    private static final long serialVersionUID = -8716257155912288696L;
 
-    public RouterMakeException() {
+    private final String routerType;
+    private final String param;
+
+    public RouterMakeException(String routerType, String param) {
+        this.routerType = routerType;
+        this.param = param;
     }
 
-    public RouterMakeException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public RouterMakeException(String message) {
-        super(message);
-    }
-
-    public RouterMakeException(Throwable cause) {
+    public RouterMakeException(Throwable cause, String routerType, String param) {
         super(cause);
+        this.routerType = routerType;
+        this.param = param;
+    }
+
+    @Override
+    public String getMessage() {
+        return "路由器构造异常, 类型为: " + routerType + ", 参数为: " + param;
     }
 }